Split GameUI component into Game and GUI
